refactor(ios): one Luke mark
`WatchLukeMark.swift` said what it was in its own second line: "Hand copy of
FaceArt + LukeMark from apps/ios/Luke/Marks.swift — change both when the
artwork moves." A copy kept in step by a comment is a copy that eventually is
not, so there is one now.

`FaceArt` and `LukeMark` move to `LukeKit`, which both apps link and which
needs no UIKit to draw them. What stays on the phone is the one thing UIKit
does: the tab bar's rasterized face, now an extension on the shared mark, and
the provider, Google, and GitHub marks, which trace SVG path data through the
app's own parser.
